Qi-364 Good move towards decoupling service activation.
Both old and new activation system work together.

Every unit test involving activation has been ported to the new system.
This except @Ignored PluginTest that rely on ability to set Service
MetaInfo at runtime, it seems it's not possible anymore?

Bonus: ImportedServices now have Activators and importOnStartup (in the
same vein as instancianteOnStartup for plain Services).

Next action is to remove ActivationEventListener that's not needed
anymore as the new activation system provide a direct way to listen to
activation events.

After that we need to remove the old activation system. For this I have
to impact every core/ext/lib Service and this will be done in a next
day because it'll be a lot of tedious work.