Removed tools inter-dependencies (Envisage & EntityViewer)
EntityViewer depended on Envisage for de detailled models only.
I extracted the detailled model related code in a module.
It's cleaner this way and will allow us to concentrate on the
model-detail module that need a test harness (I spotted some things that
should work in Envisage but don't, Service Usage for example).