Core (minor): added @Override annotation where applicable
This commit is big but contains no significant changes.

Since the codebase has history the whole core was full of methods with
or without the @Override annotation making it quite useless.

This commits adds the @Override annotation everywhere applicable so now
it's possible to distinguish methods overriding supertypes and plain
methods.

Some imports were fixed along the way.