POLYGENE-221 integration test Riak ES using Docker
Riak listen to its network port before being ready to serve clients.
This breaks the test-support integration.
The tests Thread.sleep to give it a chance to be ready
This will be flaky and should be fixed before we run these tests on CI.
See Riak ES tests for details.